Move input_event into a common input definitions header that can be shared with Cygwin and Linux build configurations.
