+++ catacurse.h
@@ -14,6 +14,9 @@
 #include <map>
 #include <vector>
 #include "json.h"
+
+#include "input_defs.h"
+
 typedef int chtype;
 typedef unsigned short attr_t;
 typedef unsigned int u_int32_t;
@@ -75,82 +78,7 @@ typedef struct {
 
 } WINDOW;
 
-enum input_event_t {
-    CATA_INPUT_ERROR,
-    CATA_INPUT_KEYBOARD,
-    CATA_INPUT_GAMEPAD,
-    CATA_INPUT_MOUSE
-};
-
 enum mouse_buttons { MOUSE_BUTTON_LEFT=1, MOUSE_BUTTON_RIGHT=2 };
-
-/**
- * An instance of an input, like a keypress etc.
- *
- * Gamepad, mouse and keyboard keypresses will be represented as `long`.
- * Whether a gamepad, mouse or keyboard was used can be checked using the
- * `type` member.
- *
- */
-struct input_event {
-    input_event_t type;
-
-    std::vector<long> modifiers; // Keys that need to be held down for
-                                 // this event to be activated.
-
-    std::vector<long> sequence; // The sequence of key or mouse events that
-                                // triggers this event. For single-key
-                                // events, simply make this of size 1.
-
-    int mouse_x, mouse_y;       // Mouse click co-ordinates, if applicable
-
-    input_event()
-    {
-        mouse_x = mouse_y = 0;
-    }
-
-    long get_first_input() const
-    {
-        if (sequence.size() == 0) {
-            return 0;
-        }
-
-        return sequence[0];
-    }
-
-    void add_input(const long input)
-    {
-        sequence.push_back(input);
-    }
-
-    bool operator==(const input_event& other) const
-    {
-        if(type != other.type) {
-            return false;
-        }
-
-        if(sequence.size() != other.sequence.size()) {
-            return false;
-        }
-        for(int i=0; i<sequence.size(); i++) {
-            if(sequence[i] != other.sequence[i]) {
-                return false;
-            }
-        }
-
-        if(modifiers.size() != other.modifiers.size()) {
-            return false;
-        }
-        for(int i=0; i<modifiers.size(); i++) {
-            if(modifiers[i] != other.modifiers[i]) {
-                return false;
-            }
-        }
-
-        return true;
-    }
-};
-
 
 #define A_NORMAL __NORMAL
 #define A_STANDOUT __STANDOUT
