+++ cursesdef.h
@@ -4,8 +4,10 @@
 #if (defined TILES || defined SDLTILES || defined _WIN32 || defined WINDOWS)
     #include "catacurse.h"
 #elif (defined __CYGWIN__)
+    #include "input_defs.h"
     #include "ncurses/curses.h"
 #else
+    #include "input_defs.h"
     #include <curses.h>
 #endif
