+++ msvc110/Cataclysm.vcxproj
@@ -993,6 +993,7 @@
     <ClInclude Include="..\iexamine.h" />
     <ClInclude Include="..\init.h" />
     <ClInclude Include="..\input.h" />
+    <ClInclude Include="..\input_defs.h" />
     <ClInclude Include="..\inventory.h" />
     <ClInclude Include="..\item.h" />
     <ClInclude Include="..\item_factory.h" />
